エラー対策

+++ Diary/app/Http/Controllers/DiaryController.php
@@ -61,6 +61,12 @@ class DiaryController extends Controller
     {
         // Diaryモデルを使用して、IDが一致する日記の取得
         $diary = Diary::find($id);
+
+        //ログインユーザーが日記の投稿者かチェックする
+        if(Auth::user()->id !=$diary->user_id){
+            //投稿者とログインユーザーが違う場合
+            abort(404);
+        }
         //取得した日記の削除
         $diary->delete();
         //一覧画面にリダイレクト
@@ -69,6 +75,12 @@ class DiaryController extends Controller
     //編集画面を表示する
     public function edit(Diary $diary)
     {
+        //ログインユーザーが日記の投稿者かチェックする
+        if(Auth::user()->id !=$diary->user_id){
+            //投稿者とログインユーザーが違う場合
+            abort(404);
+        }
+
         //受け取ったIDを元に日記を取得
         // $diary= Diary::find($id);
         //編集画面を返す。同時に画面に取得した日記を渡す。
@@ -84,6 +96,11 @@ class DiaryController extends Controller
     {
         //受けとったIDを元に日記を取得
         $diary= Diary::find($id);
+        //ログインユーザーが日記の投稿者かチェックする
+        if(Auth::user()->id !=$diary->user_id){
+            //投稿者とログインユーザーが違う場合
+            abort(404);
+        }
         //取得した日記のタイトル,本文を置き換える
         $diary->title = $request->title;
         $diary->body = $request->body;
